modesetting: 24bpp are too confusing, shadow our way out.
So we have 2 places where gpus with 24bpp frontbuffers are still
in use, cirrus (in qemu) and some early variants of the mgag200
server chips.

Currently we seem to get a lot of broken rendering in qt, mesa, gnome
if we expose the frontbuffer as 24bpp, nobody seems to test this
anymore upstream so client side apps are constantly broken, so lets
just make -modesetting expose a 24/32 shadow frontbuffer, and use
the shadow update hook to convert down to the actual 24bpp front,
this might be slower, but its correct and really anyone that wants
this already has enough problems.

Signed-off-by: Dave Airlie <airlied@redhat.com>